Move regression computations into function
